ASCII ART GENERATOR/requirements.txt
# NVIDIA Mistral client
mistralai>=0.1.11

# Image processing for ASCII art
Pillow>=10.0.0

# HTTP requests
requests>=2.30.0

# Optional: Numeric processing if needed
numpy>=1.26.0
